video_system/requirements.txt
# Core dependencies
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0
requests>=2.25.0
tqdm>=4.62.0

# Text processing and NLP
sentence-transformers>=2.2.0
transformers>=4.21.0
torch>=1.12.0
torchvision>=0.13.0

# Vector search and similarity
faiss-cpu>=1.7.0
scikit-learn>=1.0.0

# Video processing
moviepy>=1.0.3

# Azure Speech Services
azure-cognitiveservices-speech>=1.24.0

# Web scraping (for data collection)
beautifulsoup4>=4.10.0
selenium>=4.0.0

# Configuration and utilities
PyYAML>=6.0
argparse
psutil>=5.8.0

# Machine Learning and AI
huggingface-hub>=0.10.0

# Optional: GPU acceleration for FAISS
# faiss-gpu>=1.7.0  # Uncomment if you have CUDA-compatible GPU

# Development and testing
jupyter>=1.0.0
matplotlib>=3.5.0
